+++ control/src/cache/display.rs
@@ -41,6 +41,7 @@ pub enum ConfigurationScreen {
     Idle(u32),
     Rewind((usize, usize)),
     DefaultScreen(usize),
+    PositionResetMapping(Option<usize>),
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -237,6 +238,9 @@ fn ticked_dialog(menu: DialogScreen) -> DialogScreen {
             ConfigurationScreen::DefaultScreen(selection) => {
                 ticked_configuration_default_screen(selection)
             }
+            ConfigurationScreen::PositionResetMapping(mapping) => {
+                ticked_configuration_position_reset_mapping(mapping)
+            }
         },
         DialogScreen::Calibration(calibration) => match calibration {
             CalibrationScreen::SelectOctave1(i, cycles) => ticked_calibration_1(i, cycles),
@@ -257,6 +261,10 @@ fn ticked_configuration_rewind(rewind: (usize, usize)) -> DialogScreen {
 
 fn ticked_configuration_default_screen(selection: usize) -> DialogScreen {
     DialogScreen::Configuration(ConfigurationScreen::DefaultScreen(selection))
+}
+
+fn ticked_configuration_position_reset_mapping(mapping: Option<usize>) -> DialogScreen {
+    DialogScreen::Configuration(ConfigurationScreen::PositionResetMapping(mapping))
 }
 
 fn ticked_calibration_1(i: usize, mut cycles: u32) -> DialogScreen {
@@ -386,6 +394,14 @@ fn leds_for_configuration(configuration: &ConfigurationScreen) -> [bool; 8] {
             leds
         }
         ConfigurationScreen::DefaultScreen(selection) => index_to_leds(*selection),
+        ConfigurationScreen::PositionResetMapping(mapping) => {
+            let mut leds = [false; 8];
+            if let Some(index) = mapping {
+                leds[*index] = true;
+                leds[*index + 4] = true;
+            }
+            leds
+        }
     }
 }